add icons for delicious

+++ Pushpin/PPPinboardMetadataCache.m
@@ -46,11 +46,11 @@
 
 - (NSString *)cacheKeyForPost:(NSDictionary *)post compressed:(BOOL)compressed dimmed:(BOOL)dimmed width:(CGFloat)width {
     if (post[@"hash"]) {
-        return [NSString stringWithFormat:@"%@:%@:%@:%@:%d", post[@"hash"], post[@"meta"], compressed ? @"1": @"0", dimmed ? @"1" : @"0", (NSInteger)width];
+        return [NSString stringWithFormat:@"%@:%@:%@:%@:%ld", post[@"hash"], post[@"meta"], compressed ? @"1": @"0", dimmed ? @"1" : @"0", (long)width];
     }
     else {
         // It's a network item
-        return [NSString stringWithFormat:@"%@:%@:%@:%d", post[@"url"], compressed ? @"1": @"0", dimmed ? @"1" : @"0", (NSInteger)width];
+        return [NSString stringWithFormat:@"%@:%@:%@:%ld", post[@"url"], compressed ? @"1": @"0", dimmed ? @"1" : @"0", (long)width];
     }
 }
